requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
asyncio-throttle>=1.0.0
python-json-logger>=2.0.0